Rename the extension file and namespace.
